tests: cleanup
- move remaining integration test helper into integration_test.go
- where appropriate, use require.NoError(*testing.T, error)
  instead of assert.NoError(*testing.T, error), to fail a
  test immediately
