+++ util/util_test.go
@@ -10,13 +10,15 @@ func TestBytes(t *testing.T) {
 	t.Parallel()
 
 	out, err := Bytes("4160629140050E")
-	_, err2 := Bytes("4160629140050E0")
-	_, err3 := Bytes("4160629140050K")
 	exp := []byte{0x41, 0x60, 0x62, 0x91, 0x40, 0x5, 0x0e}
 	assert.NoError(t, err)
-	assert.Error(t, err2)
-	assert.Error(t, err3)
 	assert.Equal(t, exp, out)
+
+	_, err = Bytes("4160629140050E0")
+	assert.EqualError(t, err, "parse octets: uneven length of string")
+
+	_, err = Bytes("4160629140050K")
+	assert.EqualError(t, err, "parse octets: met a non-HEX rune in string")
 }
 
 func TestHexString(t *testing.T) {
